Heroku/requirements.txt
absl-py==0.7.1
alabaster==0.7.12
args==0.1.0
asn1crypto==0.24.0
astor==0.8.0
astroid==2.2.5
astropy==3.2.1
atomicwrites==1.3.0
attrs==19.1.0
Babel==2.7.0
backcall==0.1.0
backports-abc==0.5
backports.csv==1.0.7
backports.functools-lru-cache==1.5
backports.os==0.1.1
backports.shutil-get-terminal-size==1.0.0
backports.shutil-which==3.5.2
backports.tempfile==1.0
backports.weakref==1.0.post1
beautifulsoup4==4.7.1
bitarray==0.9.3
bkcharts==0.2
bleach==3.1.0
bokeh==1.2.0
boto==2.49.0
Bottleneck==1.2.1
cachetools==4.1.1
certifi==2020.6.20
cffi==1.14.1
chardet==3.0.4
click==7.1.2
client==0.0.1
clint==0.5.1
cloudpickle==1.1.1
configparser==3.7.4
contextlib2==0.5.5
cryptography==3.0
cycler==0.10.0
Cython==0.29.10
cytoolz==0.9.0.1
dask==1.2.2
decorator==4.4.2
defusedxml==0.6.0
dialogflow==1.0.0
distributed==1.28.1
docopt==0.6.2
docutils==0.14
entrypoints==0.3
enum34==1.1.6
et-xmlfile==1.0.1
extcolors==0.1.2
fastcache==1.1.0
feedparser==5.2.1
filelock==3.0.12
Flask==1.1.2
funcsigs==1.0.2
future==0.17.1
fuzzywuzzy==0.18.0
gast==0.2.2
gevent==1.4.0
glob2==0.6
gnewsclient==1.12
google-api-core==1.22.0
google-auth==1.20.1
google-pasta==0.1.7
googleapis-common-protos==1.52.0
greenlet==0.4.15
grpcio==1.31.0
gunicorn==20.0.4
h5py==2.9.0
heapdict==1.0.0
html5lib==1.0.1
idna==2.10
imageio==2.5.0
imagesize==1.1.0
importlib-metadata==0.17
ipykernel==5.1.1
ipython==7.5.0
ipython-genutils==0.2.0
ipywidgets==7.4.2
isort==4.3.20
itsdangerous==1.1.0
jdcal==1.4.1
jedi==0.13.3
Jinja2==2.11.2
joblib==0.13.2
jsonschema==3.0.1
jupyter==1.0.0
jupyter-client==5.2.4
jupyter-console==6.0.0
jupyter-core==4.4.0
jupyterlab==0.35.4
jupyterlab-launcher==0.13.1
jupyterlab-server==0.2.0
keyring==18.0.0
kiwisolver==1.1.0
lazy-object-proxy==1.4.1
libarchive-c==2.8
linecache2==1.0.0
llvmlite==0.29.0
locket==0.2.0
lxml==4.3.3
Markdown==3.1.1
MarkupSafe==1.1.1
matplotlib==3.1.0
mccabe==0.6.1
more-itertools==7.0.0
msgpack==0.6.1
multipledispatch==0.6.0
nose==1.3.7
notebook==5.7.8
numba==0.44.1
numexpr==2.6.9
numpy==1.16.4
numpydoc==0.9.1
odfpy==1.4.0
olefile==0.46
openpyxl==2.6.2
packaging==19.0
pandas==0.24.2
pandocfilters==1.4.2
parso==0.4.0
partd==0.3.10
path.py==12.0.1
pathlib2==2.3.3
patsy==0.5.1
pep8==1.7.1
pickleshare==0.7.5
Pillow==7.2.0
pkginfo==1.5.0.1
pluggy==0.12.0
ply==3.11
prometheus-client==0.6.0
prompt-toolkit==2.0.9
psutil==5.6.2
py==1.8.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.5.0
pycosat==0.6.3
pycparser==2.20
pycrypto==2.6.1
pycurl==7.43.0.2
pyflakes==2.1.1
Pygments==2.4.2
pylint==2.3.1
pyOpenSSL==19.0.0
pyparsing==2.4.0
pyreadline==2.1
pyrsistent==0.14.11
pytest==4.6.2
pytest-arraydiff==0.3
pytest-astropy==0.5.0
pytest-doctestplus==0.3.0
pytest-openfiles==0.3.2
pytest-remotedata==0.3.1
python-crontab==2.3.6
python-dateutil==2.8.0
python-telegram-bot==12.8
pytz==2020.1
PyWavelets==1.0.3
PyYAML==5.1
pyzmq==18.0.0
requests==2.24.0
rope==0.14.0
rsa==4.6
ruamel-yaml==0.15.46
simplegeneric==0.8.1
singledispatch==3.4.0.3
six==1.15.0
snowballstemmer==1.2.1
sortedcollections==1.1.2
sortedcontainers==2.1.0
soupsieve==1.8
SQLAlchemy==1.3.4
sqlparse==0.3.0
termcolor==1.1.0
terminado==0.8.2
testpath==0.4.2
toolz==0.9.0
tornado==6.0.4
tqdm==4.32.1
traceback2==1.4.0
traitlets==4.3.2
typing==3.6.4
unicodecsv==0.14.1
unittest2==1.1.0
urllib3==1.25.10
wcwidth==0.1.7
webencodings==0.5.1
webp==0.1.0a13
Werkzeug==1.0.1
widgetsnbextension==3.4.2
win-inet-pton==1.1.0
win-unicode-console==0.5
wincertstore==0.2
wrapt==1.11.1
zict==0.1.4
zipp==0.5.1